api/requirements.txt
tensorflow-cpu
flask
flask-cors
pillow
numpy
